requirements.txt
upstash-vector
openai
groq
python-dotenv